+++ mobile/src/app/(tabs)/index.tsx
@@ -1,6 +1,6 @@
 import { Image, StyleSheet } from 'react-native';
-import EditScreenInfo from '@/components/EditScreenInfo';
-import { Text, View } from '@/components/Themed';
+import EditScreenInfo from '@/src/components/EditScreenInfo';
+import { Text, View } from '@/src/components/Themed';
 import mobiLogo from '../../assets/images/mobi.png'
 
 export default function TabOneScreen() {
