+++ mobile/src/assets/images/splash.png
