+++ mobile/src/components/EditScreenInfo.tsx
@@ -6,7 +6,7 @@ import { MonoText } from './StyledText';
 import { Text, View } from './Themed';
 
 
-import Colors from '@/constants/Colors';
+import Colors from '@/src/constants/Colors';
 
 export default function EditScreenInfo({ path }: { path: string }) {
   return (
