+++ mobile/src/components/__tests__/StyledText-test.js
